+++ src/rezbuild/bin_utils.py
@@ -34,8 +34,8 @@ def change_shebang(filepath, shebang, is_bin=False, origin_shebang=""):
     if origin_shebang and not origin_shebang.startswith("#!"):
         origin_shebang = f"#!{origin_shebang}"
 
-    with open(filepath, read_mode) as file_:
-        content = file_.read()
+    with open(filepath, read_mode) as file:
+        content = file.read()
 
     if origin_shebang and is_bin:
         new_shebang = bytes(shebang, encoding="utf-8")
@@ -48,8 +48,8 @@ def change_shebang(filepath, shebang, is_bin=False, origin_shebang=""):
         origin_shebang = re.match("#!.+", content).group(0)
         new_content = content.replace(origin_shebang, shebang)
 
-    with open(filepath, write_mode) as file_:
-        file_.write(new_content)
+    with open(filepath, write_mode) as file:
+        file.write(new_content)
 
 
 def get_windows_shebang(filepath, pattern):
@@ -64,48 +64,72 @@ def get_windows_shebang(filepath, pattern):
     """
     if not pattern.startswith("#!"):
         pattern = f"#!{pattern}"
-    with open(filepath, "rb") as file_:
-        content = file_.read()
+    with open(filepath, "rb") as file:
+        content = file.read()
     match = re.search(bytes(pattern, encoding="utf-8"), content)
     if match:
         return str(match.group(0), encoding="utf-8")
     return ""
 
 
-def make_bins_movable(
-        bin_path, shebang="", pattern="", rpath="", lib_dir="",
-        change_load_dylib=True):
-    """Change all the bins in the directory to make it movable.
+def make_bin_movable(
+        bin_path, shebang="", pattern="", lib_dir="", extra_lib_dirs=None,
+        add_rpath=True):
+    """Make bin file movable.
 
     Args:
-        bin_path (str): Where the entry files placed.
-        shebang (str, optional): The shebang to change to. Default is `""`.
+        bin_path (str): Path of the bin file.
+        shebang (str, optional): The shebang to change to. Default is "".
         pattern (str, optional): The re pattern to match shebang. Default is
             `""`.
-        rpath (str, optional): The rpath to add. Default is `""`.
         lib_dir (str, optional): The destination path to put the dylib file to.
             Default is `""`.
-        change_load_dylib (str, optional): Whether to change the load dylib
-            path to rpath. Default is True.
+        extra_lib_dirs (:obj:`list` of :obj:`str`, optional): The extra library
+            directory to check if the dylib file should be copied. The dylib
+            file will not be copied when it in the exclude_lib_dirs.
+        add_rpath (bool, optional): Whether to add rpath into macho. Default is
+            True.
     """
-    for filename in os.listdir(bin_path):
-        filepath = os.path.join(bin_path, filename)
-        if platform.system() == "Windows":
-            if not pattern:
-                raise ArgumentError(
-                    "Argument `pattern` can not empty on Windows.")
-            if not shebang:
-                raise ArgumentError(
-                    "Argument `shebang` can not empty on Windows.")
-            origin_shebang = get_windows_shebang(filepath, pattern)
-            change_shebang(
-                filepath, shebang, is_bin=True, origin_shebang=origin_shebang)
+    if platform.system() == "Windows":
+        if not pattern:
+            raise ArgumentError(
+                "Argument `pattern` can not empty on Windows.")
+        origin_shebang = get_windows_shebang(bin_path, pattern)
+        change_shebang(
+            bin_path, shebang, is_bin=True, origin_shebang=origin_shebang)
+    else:
+        if MachO.is_macho(bin_path):
+            macho = MachO(bin_path)
+            if not macho.static:
+                macho.make_macho_movable(lib_dir, extra_lib_dirs, add_rpath)
         else:
-            if MachO.is_macho(filepath):
-                MachO(filepath).make_macho_movable(
-                    rpath, lib_dir, change_load_dylib)
-            else:
-                change_shebang(filepath, shebang)
+            change_shebang(bin_path, shebang)
+
+
+def make_bins_movable(
+        path, shebang="", pattern="", lib_dir="", extra_lib_dirs=None,
+        add_rpath=True):
+    """Make all the bins movable which in the directory.
+
+    Args:
+        path (str): Path to the directory.
+        shebang (str, optional): The shebang to change to. Default is "".
+        pattern (str, optional): The re pattern to match shebang. Default is
+            `""`.
+        lib_dir (str, optional): The destination path to put the dylib file to.
+            Default is `""`.
+        extra_lib_dirs (:obj:`list` of :obj:`str`, optional): The extra library
+            directory to check if the dylib file should be copied. The dylib
+            file will not be copied when it in the exclude_lib_dirs.
+        add_rpath (bool, optional): Whether to add rpath into macho. Default is
+            True.
+    """
+    for filename in os.listdir(path):
+        filepath = os.path.join(path, filename)
+        if os.path.isdir(filepath) or os.path.islink(filepath):
+            continue
+        make_bin_movable(
+            filepath, shebang, pattern, lib_dir, extra_lib_dirs, add_rpath)
 
 
 class MachO(object):
@@ -136,10 +160,33 @@ class MachO(object):
         Args:
             path (str): The path of the Mach-O file.
         """
+        self.fat = False
         self.load_dylibs = []
+        self.macho = False
         self.path = path
         self.rpaths = []
-        self.parse_load_command()
+        self.static = False
+        self._parse_file()
+
+    def _parse_file(self):
+        """Parse the MachO file."""
+        with open(self.path, "rb") as file:
+            magic_number = file.read(4)
+            if magic_number in self.MACHO_MAGIC_NUMBERS:
+                self.macho = True
+            else:
+                return
+            if magic_number == b"\xca\xfe\xba\xbe":
+                self.fat = True
+                arch_count = struct.unpack(">I", file.read(4))[0]
+                # Each fat header has 5 entity. Each entity occupies 4 bytes.
+                offsets = self.get_fat_offsets(
+                    arch_count, file.read(20 * arch_count))
+                content = b"\x00" * (8 + arch_count * 20) + file.read()
+            else:
+                offsets = [0]
+                content = magic_number + file.read()
+        self.parse_arch(content, offsets)
 
     def add_rpath(self, rpath):
         """Add rpath into Macho-O file.
@@ -148,6 +195,15 @@ class MachO(object):
             rpath (str): The rpath to add.
         """
         cmd = ["install_name_tool", "-add_rpath", rpath, self.path]
+        subprocess.run(cmd, check=True)
+
+    def change_dylib_id(self, dylib_id):
+        """change the load dylib ID.
+
+        Args:
+            dylib_id (str): The id to change to.
+        """
+        cmd = ["install_name_tool", "-id", dylib_id, self.path]
         subprocess.run(cmd, check=True)
 
     def change_load_dylib(self, old_path, new_path):
@@ -198,6 +254,20 @@ class MachO(object):
             "@executable_path",
             get_relative_path(os.path.dirname(self.path), libdir))
 
+    @staticmethod
+    def get_fat_offsets(arch_count, content):
+        """Get the offsets from the MachO universal binary file.
+
+        Args:
+            arch_count (int): The architecture number of the MachO file.
+            content (bytes): The header content of the fat file.
+        """
+        offsets = []
+        for i in range(arch_count):
+            offset = struct.unpack(">I", content[i * 20 + 8:i * 20 + 12])[0]
+            offsets.append(offset)
+        return offsets
+
     def get_load_dylibs(self, ignore_system_dylib=False):
         """Get all the dylibs in this Mach-O files.
 
@@ -209,6 +279,16 @@ class MachO(object):
             return self.get_non_system_dylibs()
         else:
             return self.load_dylibs
+
+    @staticmethod
+    def get_magic_number(filepath):
+        """Get magic number from file.
+
+        Args:
+            filepath (str): The file to get magic number from.
+        """
+        with open(filepath, "rb") as f:
+            return f.read(4)
 
     def get_non_system_dylibs(self):
         """Get dylibs that doesn't contain system dylibs."""
@@ -231,29 +311,31 @@ class MachO(object):
         Returns:
             bool: True if the file is a Mach-O file, false otherwise.
         """
-        with open(filepath, "rb") as f:
-            magic_number = f.read(4)
-        return magic_number in cls.MACHO_MAGIC_NUMBERS
+        return cls.get_magic_number(filepath) in cls.MACHO_MAGIC_NUMBERS
 
-    def make_macho_movable(self, rpath="", lib_dir="", change_load_dylib=True):
+    def make_macho_movable(
+            self, lib_dir="", extra_lib_dirs=None, add_rpath=True):
         """Make the macho file movable.
 
-        Copy the unmovable dylibs into the dst_dir and add the rpath into
-        macho.
-
         Args:
-            rpath (str, optional): The rpath to add. Default is the relative
-                path between the macho_path and dst_dir.
             lib_dir (str, optional): The destination path to put the dylib file
                 to. Default is the "lib" folder in the same level directory as
                 the parent directory of macho_path.
-            change_load_dylib (bool, optional): Whether to change the load
-                dylib path to rpath. Default is True.
+            extra_lib_dirs (:obj:`list` of :obj:`str`, optional): The extra
+                library directory to add to rpath. The dylib files will not be
+                copied when it already in extra_lib_dirs.
+            add_rpath (bool, optional): Whether to add rpath into macho.
+                Default is True.
         """
         lib_dir = lib_dir or self.get_default_libdir()
-        rpath = rpath or self.get_default_rpath(lib_dir)
-        if change_load_dylib and rpath not in self.rpaths:
-            self.add_rpath(rpath)
+        extra_lib_dirs = extra_lib_dirs or []
+        extra_lib_dirs.append(lib_dir)
+        if add_rpath:
+            for dir_ in extra_lib_dirs:
+                rpath = self.get_default_rpath(dir_)
+                if rpath not in self.rpaths:
+                    self.add_rpath(rpath)
+                    self.rpaths.append(rpath)
 
         def copy_and_change(macho):
             """Copy dylibs and change lcload path.
@@ -261,30 +343,46 @@ class MachO(object):
             Args:
                 macho (MachO): The macho object to change.
             """
+            macho.change_dylib_id(f"@rpath/{os.path.basename(macho.path)}")
             for dylib in macho.get_load_dylibs(ignore_system_dylib=True):
                 lib_name = os.path.basename(dylib)
-                if change_load_dylib:
-                    macho.change_load_dylib(dylib, f"@rpath/{lib_name}")
-                if lib_name not in os.listdir(lib_dir):
+                macho.change_load_dylib(dylib, f"@rpath/{lib_name}")
+                for extra_lib_dir in extra_lib_dirs:
+                    if lib_name in os.listdir(extra_lib_dir):
+                        copy_and_change(
+                            MachO(os.path.join(extra_lib_dir, lib_name)))
+                        break
+                else:
                     dst = os.path.join(lib_dir, lib_name)
                     shutil.copy2(dylib, dst)
-                    copy_and_change(MachO(dylib))
+                    copy_and_change(MachO(dst))
 
         copy_and_change(self)
 
-    def parse_load_command(self):
-        """Parse Mach-O file load commands."""
-        with open(self.path, "rb") as file_:
-            file_.read(16)
-            load_cmd_count = struct.unpack("I", file_.read(4))[0]
-            file_.read(12)
+    def parse_arch(self, content, offsets):
+        """Parse the MachO file architecture.
+
+        Args:
+            content (bytes): The MachO file content.
+            offsets (:obj:`list` of :obj:`int`): The offset of each
+                architecture in MachO file.
+        """
+        for offset in offsets:
+            magic_number = content[offset:offset + 4]
+            if magic_number not in self.MACHO_MAGIC_NUMBERS:
+                self.static = True
+                break
+            load_cmd_count = struct.unpack(
+                "I", content[offset + 16:offset + 20])[0]
+            feed = 32 + offset
             for _ in range(load_cmd_count):
-                cmd, size = struct.unpack("2I", file_.read(8))
-                content = file_.read(size - 8)
+                cmd, size = struct.unpack("2I", content[feed:feed + 8])
                 if cmd not in self.LOAD_COMMANDS_MAPPING.keys():
+                    feed += size
                     continue
                 func = self.LOAD_COMMANDS_MAPPING[cmd]
-                getattr(self, func)(content)
+                getattr(self, func)(content[feed + 8:feed + size])
+                feed += size
 
     def parse_rpath(self, content):
         """Parse rpath from the given rpath content.